+++ CHANGES.md
@@ -1,6 +1,9 @@
 <!-- vim:ts=4:sts=4:sw=4:et:tw=60 -->
 
 ## 0.9.6
+  - Introdues `series` for composing a linear workflow (chaining functions on a promise)
+
+## 0.9.6
 
  - Fixes a bug in recognizing the difference between compatible Q
    promises, and Q promises from before the implementation of "inspect".
@@ -37,7 +40,7 @@
 
 ## 0.9.4
 
- - `isPromise` and `isPromiseAlike` now always returns a boolean 
+ - `isPromise` and `isPromiseAlike` now always returns a boolean
    (even for falsy values). #284 @lfac-pt
  - Support for ES6 Generators in `async` #288 @andywingo
  - Clear duplicate promise rejections from dispatch methods #238 @SLaks